% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRGpostProcess.R
\name{MRGpostProcess}
\alias{MRGpostProcess}
\title{Make some final adjustments to the multiresolution grids}
\usage{
MRGpostProcess(himg, vars, remCols = TRUE, rounding = "varying")
}
\arguments{
\item{himg}{A multi-resolution grid, typically resulting from a call to multiResGrid}

\item{vars}{Variable(s) of interest that should be aggregated (necessary when ifg is
used for individual farm specific anonymization rules)}

\item{remCols}{Logical; Should intermediate columns be removed? Can be set
to FALSE for further analyses. Temporary columns will not be removed if their names 
partly match the variable names of \code{vars}}

\item{rounding}{Either 'FALSE', an integer specifying the number of decimal
                           places to use, or `"varying"`. Negative integers are allowed (for example,
                           `-1` rounds to the nearest 10); see also the `digits` argument in
                          \code{\link{round}}.

                           If `rounding = "varying"`, values are rounded according to their leading
                           digit, following the current default used in Eurostat layers. Most values
                           are rounded to **one significant digit** (for example, 788 becomes 800 and
                           53 becomes 50). However, if the first significant digit is 1 or 2, values
                           are rounded to **two significant digits** instead (for example, 117 becomes
                           120 and 287 becomes 290).}
}
\value{
The function will return a post-processed multi-resolution grid
with non-confidential gridded data. See \code{\link{multiResGrid}} for more 
information.
}
\description{
Make some final adjustments to the multiresolution grids
}
\details{
The postprocessing function is normally called directly from \code{\link{multiResGrid}}. 
However, it might be useful to check the values of the grid cells that
will be suppressed, and the values before rounding. In that case 
\code{\link{multiResGrid}} can be called with the argument \code{postProcess = FALSE}, 
and the post processing be done separately.
}
\examples{
\donttest{
library(sf)

# These are SYNTHETIC agricultural FSS data 
data(ifs_dk) # Census data
# Create spatial data
ifg = fssgeo(ifs_dk, locAdj = "LL")

# Set the base resolutions, and create a hierarchical list with gridded data
ress = 1000*2^(1:7)
ifl = gridData(ifg, "UAA", res = ress)
himg = multiResGrid(ifl, ifg = ifg, var = "UAA", weight = "EXT_CORE", postProcess = FALSE)
himgp = MRGpostProcess(himg, var = "UAA")

# Confidential grid cells, being suppressed in postProcessing
himg[himg$confidential,]
 }
}
